+++ RELEASES.md
@@ -5,7 +5,7 @@ Released on TBD (UTC).
 ### Enhancements
 - Added Bybit base coin for fee rate parsing (#1696), thanks @filipmacek
 - Updated installation from source docs for Clang dependency (#1690), thanks @Troubladore
-- Updated `InteractiveBrokersGatewayConfig` docs (#1691), thanks @Troubladore
+- Updated `DockerizedIBGatewayConfig` docs (#1691), thanks @Troubladore
 
 ### Breaking Changes
 None